mqtt_validation_project _V2/requirements.txt
paho-mqtt>=1.6.1
pandas>=2.0.0
pandera>=0.14.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
